pyproject.toml
[project]
name = "thatrack"
version = "0.1.0"
description = "an application for tracking study and work time"
authors = [
    {name = "thediogen",email = "evgendvodnenko@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "greenlet (>=3.1.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
